+++ packages/bloc_error_control_generator/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.2
+- Changed ErrorStateFor annotation - now generic
+- Fixed dependencies
+- Updated example project
+- Updated README
+
 ## 1.1.1
 - fixing dependencies